pyproject.toml
[tool.poetry]
name = "dbf2sql-uploader"
version = "0.1.0"
description = "DBF to MS SQL Server upload data provider"
authors = ["Dmitriy Grigorev <dimag@inbox.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
pyodbc = "^4.0.32"
pymssql = "^2.2.5"
pydantic = "^1.9.1"
python-dotenv = "^0.20.0"
Django = "4.0.6"
mssql-django = "^1.1.3"
pdoc3 = "^0.10.0"
django-import-export = "^2.8.0"
django-environ = "^0.9.0"
eventlet = "^0.33.1"
gevent = "^21.12.0"
dramatiq = {extras = ["redis"], version = "^1.13.0"}
dj-database-url = "^0.5.0"
django-dramatiq = "^0.11.0"
APScheduler = "^3.9.1"
watch = "^0.2.7"
pytz = "^2022.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
pre-commit = "^2.17.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
	\.git
| \.hg
| \.mypy_cache
| \.tox
| \.venv
| _build
| buck-out
| build
)/
'''